+++ eo_man/view/tool_bar.py
@@ -27,13 +27,13 @@ class ToolBar():
 
         # placed at the right end
         b = DonationButton(f, relief=GROOVE, small_icon=True).pack(side=RIGHT, padx=(0,2), pady=2)
-        b = self._create_img_button(f, "GitHub.com: EnOcean Device Manager Repository", ImageGallery.get_github_icon(), menu_presenter.open_eo_man_repo)
+        b = self._create_img_button(f, "GitHub: EnOcean Device Manager Repository", ImageGallery.get_github_icon(), menu_presenter.open_eo_man_repo)
         b.pack(side=RIGHT, padx=(0,2), pady=2)
-        b = self._create_img_button(f, "GitHub.com: EnOcean Device Manager Documentation", ImageGallery.get_help_icon(), menu_presenter.open_eo_man_documentation)
+        b = self._create_img_button(f, "GitHub: EnOcean Device Manager Documentation", ImageGallery.get_help_icon(), menu_presenter.open_eo_man_documentation)
         b.pack(side=RIGHT, padx=(0,2), pady=2)
 
     def _create_img_button(self, f:Frame, tooltip:str, image:ImageTk.PhotoImage, command) -> Button:
-        b = Button(f, image=image, relief=GROOVE, command=command)
+        b = Button(f, image=image, relief=GROOVE, cursor="hand2", command=command)
         Hovertip(b,tooltip,300)
         b.image = image
         b.pack(side=LEFT, padx=(2,0), pady=2)